# frontend/config/swagger.yaml
swagger: '2.0'
info:
  description: 'This is the theme parks API spec.'
  version: '1.0.0'
  title: 'Themepark API'
  termsOfService: 'http://swagger.io/terms/'
  license:
    name: 'Apache 2.0'
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: 'com.themeparks.api'
...
schemes:
- 'https'
- 'http'
paths:
...
  /user:
    get:
      tags:
      - 'user'
      summary: 'Get user by user name'
      description: ''
      operationId: 'getUserByName'
      produces:
      - 'application/xml'
      - 'application/json'
      parameters:
      - name: 'userId'
        in: 'query'
        description: 'The name that needs to be fetched. Use user1 for testing. '
        required: true
        type: 'string'
      responses:
        '200':
          description: 'successful operation'
          schema:
            $ref: '#/definitions/CurrentUser'
        '400':
          description: 'Invalid userId supplied'
        '404':
          description: 'User not found'
    put:
      tags:
      - 'user'
      summary: 'Update user'
      description: 'This can only be done by the logged in user.'
      operationId: 'updateUser'
      produces:
      - 'application/xml'
      - 'application/json'
      parameters:
      - name: 'userId'
        in: 'query'
        description: 'name that need to be updated'
        required: true
        type: 'string'
      - in: 'body'
        name: 'body'
        description: 'Updated user object'
        required: true
        schema:
          $ref: '#/definitions/User'
      responses:
        '200':
          description: 'successful operation'
          schema:
            $ref: '#/definitions/User'
        '400':
          description: 'Invalid user supplied'
        '404':
          description: 'User not found'
    delete:
      tags:
      - 'user'
      summary: 'Delete user'
      description: 'This can only be done by the logged in user.'
      operationId: 'deleteUser'
      produces:
      - 'application/xml'
      - 'application/json'
      parameters:
      - name: 'userId'
        in: 'query'
        description: 'The name that needs to be deleted'
        required: true
        type: 'string'
      responses:
        '400':
          description: 'Invalid userId supplied'
        '404':
          description: 'User not found'
  /user/group:
    get:
      tags:
      - 'user'
      summary: 'Get group of the current user'
      description: ''
      operationId: 'getUserGroup'
      produces:
      - 'application/xml'
      - 'application/json'
      parameters:
      - name: 'userId'
        in: 'query'
        description: 'user account ID that owns the group'
        required: true
        type: 'string'
      responses:
        '200':
          description: 'successful operation'
          schema:
            $ref: '#/definitions/Users'
    put:
      tags:
      - 'user'
      summary: 'AddUserToGroup'
      description: 'This can only be done by the logged in user.'
      operationId: 'addUserToGroup'
      produces:
      - 'application/xml'
      - 'application/json'
      parameters:
      - name: 'userId'
        in: 'query'
        description: 'user account ID that owns the group'
        required: true
        type: 'string'
      - in: 'body'
        name: 'body'
        description: 'user that needs to add to group'
        required: true
        schema:
          $ref: '#/definitions/User'
      responses:
        '200':
          description: 'successful operation'
          schema:
            $ref: '#/definitions/Users'
        '400':
          description: 'Invalid user supplied'
        '404':
          description: 'User not found'
    delete:
      tags:
      - 'user'
      summary: 'Delete user'
      description: 'This can only be done by the logged in user.'
      operationId: 'deletUserFromGroup'
      produces:
      - 'application/xml'
      - 'application/json'
      parameters:
      - name: 'userId'
        in: 'query'
        description: 'user account ID that owns the group'
        required: true
        type: 'string'
      - name: 'userToDelete'
        in: 'query'
        description: 'The user that needs to be deleted from the group'
        required: true
        type: 'string'
      responses:
        '200':
          description: 'successful operation'
        '400':
          description: 'Invalid userId supplied'
        '404':
          description: 'User not found'
definitions:
  Users:
    type: array
    items:
      $ref: '#/definitions/User'
  CurrentUser:
    type: object
    properties:
      id:
        type: 'integer'
        format: 'int64'
      firstName:
        type: 'string'
      lastName:
        type: 'string'
      email:
        type: 'string'
      password:
        type: 'string'
      phone:
        type: 'string'
      country:
        type: string
      access:
        type: string
  User:
    type: 'object'
    properties:
      id:
        type: 'integer'
        format: 'int64'
      firstName:
        type: 'string'
      lastName:
        type: 'string'
      email:
        type: 'string'
      password:
        type: 'string'
      phone:
        type: 'string'
...